fix weird behavior when the template has already been created

+++ packages/create-email/src/index.js
@@ -9,8 +9,6 @@ import ora from 'ora';
 import { tree } from './tree.js';
 
 const init = async (name) => {
-  const spinner = ora('Preparing files...\n').start();
-
   let projectPath = name;
 
   if (!projectPath) {
@@ -24,6 +22,13 @@ const init = async (name) => {
   const __dirname = path.dirname(fileURLToPath(import.meta.url));
   const templatePath = path.resolve(__dirname, '../template');
   const resolvedProjectPath = path.resolve(projectPath);
+
+  if (fse.existsSync(resolvedProjectPath)) {
+    console.error(`Project called ${projectPath} already exists!`);
+    process.exit(1);
+  }
+
+  const spinner = ora('Preparing files...\n').start();
 
   fse.copySync(templatePath, resolvedProjectPath, {
     recursive: true,
